refactored symmetry. symmetry->structure
