+++ README.md
@@ -1,5 +1,5 @@
 <div align="center">
-  <img width="300" src="docs/source/_static/logo_dark.png#gh-dark-mode-only">
+  <img width="200" src="docs/source/_static/logo_dark.png#gh-dark-mode-only">
   <img width="300" src="docs/source/_static/logo.png#gh-light-mode-only">
 </div>
